# api-specification.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Organix Backend API",
    "description": "API RESTful para sistema de gestão de conteúdo multi-tenant",
    "contact": { "name": "Organix Team", "email": "contact@organix.com" },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "Generated server url" }
  ],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/api/drafts/{id}": {
      "get": {
        "tags": ["Drafts"],
        "summary": "Buscar rascunho por ID",
        "description": "Busca um rascunho específico por ID",
        "operationId": "getDraftById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do rascunho",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Rascunho encontrado",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/DraftResponse" }
              }
            }
          },
          "404": {
            "description": "Rascunho não encontrado",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/DraftResponse" }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      },
      "put": {
        "tags": ["Drafts"],
        "summary": "Atualizar rascunho",
        "description": "Atualiza os dados de um rascunho existente",
        "operationId": "updateDraft",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do rascunho",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateDraftRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Rascunho atualizado com sucesso",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/DraftResponse" }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/DraftResponse" }
              }
            }
          },
          "404": {
            "description": "Rascunho não encontrado",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/DraftResponse" }
              }
            }
          },
          "403": {
            "description": "Sem permissão para atualizar este rascunho",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/DraftResponse" }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      },
      "delete": {
        "tags": ["Drafts"],
        "summary": "Excluir rascunho",
        "description": "Exclui um rascunho",
        "operationId": "deleteDraft",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do rascunho",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "404": { "description": "Rascunho não encontrado" },
          "204": { "description": "Rascunho excluído com sucesso" },
          "403": { "description": "Sem permissão para excluir este rascunho" }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
...
    "/api/drafts": {
      "get": {
        "tags": ["Drafts"],
        "summary": "Listar rascunhos",
        "description": "Lista rascunhos do usuário ou todos (se ADMIN)",
        "operationId": "getAllDrafts",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filtrar por status",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "productId",
            "in": "query",
            "description": "Filtrar por produto",
            "required": false,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de rascunhos retornada com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/DraftResponse" }
                }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      },
      "post": {
        "tags": ["Drafts"],
        "summary": "Criar rascunho",
        "description": "Cria um novo rascunho",
        "operationId": "createDraft",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateDraftRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Dados inválidos",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/DraftResponse" }
              }
            }
          },
          "201": {
            "description": "Rascunho criado com sucesso",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/DraftResponse" }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/drafts/{id}/approve": {
      "post": {
        "tags": ["Drafts"],
        "summary": "Aprovar rascunho",
        "description": "Aprova um rascunho para conversão em conteúdo",
        "operationId": "approveDraft",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do rascunho",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "403": {
            "description": "Acesso negado - apenas administradores",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/DraftResponse" }
              }
            }
          },
          "200": {
            "description": "Rascunho aprovado com sucesso",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/DraftResponse" }
              }
            }
          },
          "404": {
            "description": "Rascunho não encontrado",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/DraftResponse" }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
...
    "/api/drafts/{id}/status": {
      "patch": {
        "tags": ["Drafts"],
        "summary": "Alterar status do rascunho",
        "description": "Altera o status de um rascunho (apenas administradores)",
        "operationId": "updateDraftStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do rascunho",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Novo status",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "403": {
            "description": "Acesso negado - apenas administradores",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/DraftResponse" }
              }
            }
          },
          "200": {
            "description": "Status alterado com sucesso",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/DraftResponse" }
              }
            }
          },
          "404": {
            "description": "Rascunho não encontrado",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/DraftResponse" }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UpdateDraftRequest": {
        "required": ["title"],
        "type": "object",
        "properties": {
          "title": { "maxLength": 200, "minLength": 0, "type": "string" },
          "content": { "maxLength": 5000, "minLength": 0, "type": "string" },
          "status": {
            "type": "string",
            "enum": ["DRAFT", "REVIEW", "APPROVED", "REJECTED"]
          }
        },
        "description": "Novos dados do rascunho"
      },
      "DraftResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "content": { "type": "string" },
          "status": {
            "type": "string",
            "enum": ["DRAFT", "REVIEW", "APPROVED", "REJECTED"]
          },
          "productId": { "type": "string", "format": "uuid" },
          "productName": { "type": "string" },
          "userId": { "type": "string", "format": "uuid" },
          "userName": { "type": "string" },
          "companyId": { "type": "string", "format": "uuid" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "CreateDraftRequest": {
        "required": ["productId", "title"],
        "type": "object",
        "properties": {
          "title": { "maxLength": 200, "minLength": 0, "type": "string" },
          "content": { "maxLength": 5000, "minLength": 0, "type": "string" },
          "productId": { "type": "string", "format": "uuid" }
        },
        "description": "Dados do novo rascunho"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT Token para autenticação. Formato: Bearer {token}",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}